Remove store modules. Fix setting login progress flag.
