+++ frontend/src/store/actions.js
@@ -1,7 +1,8 @@
-import { getMyself } from '../../services/api/user';
+import { login, saveToken } from '../services/api/auth';
+import { getMyself } from '../services/api/user';
+
 import * as mutations from './mutations';
 
-import { login, saveToken } from '../../services/api/auth';
 
 export const FETCH_USER = 'FETCH_USER';
 export const LOGIN = 'LOGIN';
@@ -12,12 +13,16 @@ export const actions = {
       commit(mutations.SET_USER, { user });
     });
   },
-  [LOGIN]({ dispatch }, { username, password }) {
+  [LOGIN]({ commit, dispatch }, { username, password }) {
+    commit(mutations.SET_LOGIN_PROGRESS, true);
     login(username, password)
       .then((token) => {
         saveToken(token);
         dispatch(FETCH_USER);
       })
-      .catch(e => console.log(e));
+      .catch(e => console.log(e))
+      .finally(() => {
+        commit(mutations.SET_LOGIN_PROGRESS, false);
+      });
   },
 };
