+++ frontend/src/store/index.js
@@ -1,15 +1,22 @@
 import Vue from 'vue';
 import Vuex from 'vuex';
 
-import user from './modules/user';
+import { rehydrateUser } from './utils';
+import { actions } from './actions';
+import { mutations } from './mutations';
 
 const debug = process.env.NODE_ENV !== 'production';
 
 Vue.use(Vuex);
 
+const state = {
+  user: rehydrateUser(),
+  loginInProgress: false,
+};
+
 export default new Vuex.Store({
-  modules: {
-    user,
-  },
   strict: debug,
+  state,
+  actions,
+  mutations,
 });
