--- frontend/src/store/modules/mutations.js
@@ -1,7 +1,0 @@
-export const SET_USER = 'SET_USER';
-
-export const mutations = {
-  [SET_USER](state, { user }) {
-    Object.assign(state, { user });
-  },
-};
